fix/separate-api-interfaces
